fix(formatting): better support for RTL languages
Alpha support for the unified Intl API was causing improper placement of
the minus sign for RTL languages.
Now, implement of the unified Intl API is only called when an option
specifc to that API is used.
There still is improper placement of the plus sign for RTL languages but
it only affects alpha functionality.